requirements.txt
google-generativeai>=0.8.0
google-adk>=0.1.0
